requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-settings==2.8.1
alembic==1.15.2
python-multipart==0.0.20
reportlab==4.4.3
qrcode==8.2
